--- app/upgrades/v043/upgrades.go
@@ -1,34 +1,0 @@
-package v043
-
-import (
-	"context"
-
-	storetypes "cosmossdk.io/store/types"
-	upgradetypes "cosmossdk.io/x/upgrade/types"
-
-	"github.com/cosmos/cosmos-sdk/types/module"
-
-	"github.com/CosmWasm/wasmd/app/upgrades"
-)
-
-// UpgradeName defines the on-chain upgrade name
-const UpgradeName = "v0.43"
-
-var Upgrade = upgrades.Upgrade{
-	UpgradeName:          UpgradeName,
-	CreateUpgradeHandler: CreateUpgradeHandler,
-	StoreUpgrades: storetypes.StoreUpgrades{
-		Added:   []string{},
-		Deleted: []string{},
-	},
-}
-
-func CreateUpgradeHandler(
-	mm *module.Manager,
-	configurator module.Configurator,
-	ak *upgrades.AppKeepers,
-) upgradetypes.UpgradeHandler {
-	return func(ctx context.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
-		return mm.RunMigrations(ctx, configurator, fromVM)
-	}
-}
